backend/foodgram/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
django-filter==21.1
djoser==2.1.0
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==6.0.0
idna==3.4
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
oauthlib==3.2.2
Pillow==10.0.1
psycopg2-binary==2.9.9
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.5
webcolors==1.13
gunicorn==20.1.0